requirements.txt
docker-py
google-cloud
